+++ appcenter-post-build.sh
@@ -1,9 +1,6 @@
 #!/usr/bin/env bash
 
-echo "Build ID: $APPCENTER_BUILD_ID"
-echo "Branch: $APPCENTER_BRANCH"
-
-if [ "$APPCENTER_BRANCH" == "deploy-stage"  ] || [ "$APPCENTER_BRANCH" == "production" ]
+if [ "$APPCENTER_BRANCH" == "deploy-stage" ] || [ "$APPCENTER_BRANCH" == "production" ]
 then
   yarn bugsnag
 fi