+++ CHANGELOG.md
@@ -4,6 +4,19 @@ Also: [Deployment history](https://github.com/hippware/tr-wiki/wiki/Client-deplo
 
 Ticket numbers refer to the ticket tracker for this project if not specified.
 
+* UI and refactoring for #3584, #3585 (PR #3595)
+  * [android] [dev] Reconnection for disabled location settings (#3584)
+  * [android] open app location settings (#3585)
+
+# 4.5.0 - 2019 Apr 4
+
+* [android] Open app location settings (#3585)
+  * [android] Open Settings does not work if user taps "Deny" (#3561)
+* [android] fix rn-blur issue, double pop-ups (#3447, PR #3583)
+* Update: CodePush, react-native-background-geolocation (PR #3592)
+
+# 4.4.0 - 2019 Apr 2
+
 * Don't cancel location shares upon being invisible (#3556)
 * Long Press and then the X and Next buttons do not function (#3535)
 * [dev] Use version, os to identify sourcemaps (Rework #3498, PR #3568)